Fixes to get things running on 2.4.3.  - iozone runs very well  - copying large trees fails when debugging is enabled

+++ lustre/demos/baseclean.sh
@@ -33,7 +33,7 @@ if [ "$LOOPDEV" ]; then
     rmmod loop > /dev/null 2>&1
 fi
 
-if [ "$TMPFILE" -a -f "$TMPFILE" ]; then
+if [ "$LOOPDEV" -a "$TMPFILE" -a -f "$TMPFILE" ]; then
     rm -i $TMPFILE
 fi
